test(cli): consolidate ordering and notification contracts
