+++ tests/cli/test_markdown_event_order.py
@@ -1,6 +1,7 @@
 """Worker deltas and tool notifications share the real prompt-toolkit output queue."""
 import asyncio
 
+import pytest
 from prompt_toolkit.formatted_text import fragment_list_to_text, to_formatted_text
 
 from cli import HermesCLI
@@ -61,4 +62,37 @@ def test_worker_markdown_boundary_and_tool_output_keep_event_order(monkeypatch, 
             assert not cli._markdown_stream.pending
             assert tui.editor.text == 'draft typed during streaming'
 
-    asyncio.run(asyncio.wait_for(exercise(), timeout=10))
+
+@pytest.mark.parametrize('mode', ['render', 'raw', 'strip'])
+@pytest.mark.parametrize('completed_only', [False, True])
+def test_codex_tools_keep_assistant_messages_in_event_order(monkeypatch, completed_only, mode,
+                                                           codex_bridge):
+    import cli as facade
+    emitted = []
+    monkeypatch.setattr(facade, '_cprint', emitted.append)
+    cli = HermesCLI.__new__(HermesCLI)
+    cli.show_reasoning = False
+    cli.final_response_markdown = mode
+    cli.show_timestamps = False
+    cli.tool_progress_mode = 'all'
+    cli._pending_tool_info = {}
+    cli._last_scrollback_tool = None
+    cli._turn_summary_record = lambda *args: None
+    cli._invalidate = lambda *args, **kwargs: None
+    cli._reset_stream_state()
+    bridge = codex_bridge(cli)
+    emitted.append('USER: Read my note')
+    bridge({'method': 'item/agentMessage/delta', 'params': {'delta': "I'll find your note."}})
+    item = {'id': 'tool-one', 'type': 'commandExecution', 'command': 'cat note.md',
+            'aggregatedOutput': 'Today I wrote a regression test.', 'exitCode': 0, 'durationMs': 10}
+    if not completed_only:
+        bridge({'method': 'item/started', 'params': {'item': item}})
+        assert "I'll find your note." in '\n'.join(emitted)
+        assert cli._markdown_stream is None
+    bridge({'method': 'item/completed', 'params': {'item': item}})
+    bridge({'method': 'item/agentMessage/delta', 'params': {'delta': 'Here is your note.'}})
+    cli._flush_stream()
+    output = '\n'.join(emitted)
+    assert output.index('USER:') < output.index("I'll find") < output.index('exec_comm') < output.index('Here is')
+    assert output.count("I'll find your note.") == output.count('Here is your note.') == 1
+    assert not any("I'll find" in entry and 'Here is' in entry for entry in emitted)
